Fix percision lost when calculating the frame time. Started after ~4 minutes due to float percision limits, would be visible after ~18-36 hours when it would be off by 4-8ms.

+++ src/d2dx/RenderContext.h
@@ -213,7 +213,7 @@ namespace d2dx
 		int64_t _timeStart;
 		bool _hasAdjustedWindowPlacement = false;
 
-		double _prevTime;
+		int64_t _prevTime;
 		double _frameTimeMs;
 	};
 }
